fixed file delete feature

+++ app/Http/Controllers/Admin/SettingsController.php
@@ -50,7 +50,7 @@ class SettingsController extends Controller
             if($file) {
                 foreach($file as $key => $f) {      
                     if($f) {
-                        if(file_exists(public_path($settings->$key))){
+                        if(file_exists(public_path($settings->$key)) && $settings->key){
                             unlink(public_path($settings->$key));
                         }
                         $fileName = time().'_'.$f->getClientOriginalName();
